project/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-webpack-loader==1.1.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
